% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindall.R
\name{bindall}
\alias{bindall}
\title{Bindall: binding all particular type datasets in a repository}
\usage{
bindall(repo, type = "xlsx")
}
\arguments{
\item{repo}{A string with the name and/or destination of the repository}

\item{type}{The targeted type of the data files.
Currently support XLSX ("xlsx"), ODS ("ods"), and CSV ("csv").}
}
\value{
A dataframe containing the binded datasets of all
files of the requested type in the repository.
}
\description{
A function for binding all dataset files in a
particular repository and writing them into a new dataframe.
Current supported input format: CSV, XLSX, and ODS
}
\examples{
df <- bindall(repo = "Examples/XLSX/", type = "xlsx")
}
