% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databinders.R
\docType{package}
\name{databinders}
\alias{databinders}
\alias{databinders-package}
\title{Binding Data Made Easy}
\description{
The aim of databinders is to solve a common problem
of mine: the need for
binding many datasets containing the experimental data of
my many specimens and recapitulating them into a single dataframe.
Note that this package was basically written as a wrapper of
the well-known \href{https://cran.r-project.org/web/packages/rio/}{rio} package,
my preferred go-to package for working with many types of data.
}
\seealso{
\code{\link{bindall}}
}
\author{
\strong{Maintainer:} Ahmad Alkadri (email: \href{mailto:ahmad.alkadri@outlook.com}{ahmad.alkadri@outlook.com})
}
